+++ StarkBank/StarkBankTests/User.cs
@@ -13,7 +13,7 @@ namespace StarkBankTests
                 Environment.GetEnvironmentVariable("SANDBOX_ID"), // "9999999999999999",
                 Environment.GetEnvironmentVariable("SANDBOX_PRIVATE_KEY") // "-----BEGIN EC PRIVATE KEY-----\nMHQCAQEEIBEcEJZLk/DyuXVsEjz0w4vrE7plPXhQxODvcG1Jc0WToAcGBSuBBAAK\noUQDQgAE6t4OGx1XYktOzH/7HV6FBukxq0Xs2As6oeN6re1Ttso2fwrh5BJXDq75\nmSYHeclthCRgU8zl6H1lFQ4BKZ5RCQ==\n-----END EC PRIVATE KEY-----\n"
             );
-            User.Default = project;
+            Settings.User = project;
             return project;
         }
     }